requirements.txt
absl-py==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
base58==1.0.3
bcrypt==3.2.0
bleak==0.22.2
blinker==1.4
cbor==1.0.0
certifi==2024.7.4
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
Cython==0.29.28
dbus-fast==2.22.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
ecdsa==0.18.0b1
flatbuffers==24.3.25
fonttools==4.53.1
future==1.0.0
gast==0.6.0
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.65.5
h5py==3.11.0
hid==1.0.6
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iso8601==2.1.0
jeepney==0.7.1
keras==3.5.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lz4==3.1.3+dfsg
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mnemonic==0.19
more-itertools==8.10.0
MouseInfo==0.1.3
mpi4py==3.1.3
msgpack==1.0.3
mslex==1.2.0
namex==0.0.8
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
oslex==0.1.3
packaging==21.3
passlib==1.7.4
pbr==5.8.0
pillow==10.4.0
ply==3.11
protobuf==4.25.4
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.9.0
pypng==0.0.20
PyQRCode==1.2.1
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-snappy==0.5.3
python3-xlib==0.15
PyTrie==0.4.0
pytweening==1.2.0
PyYAML==5.4.1
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
SecretStorage==3.3.1
serial==0.0.97
service-identity==18.1.0
six==1.16.0
sortedcontainers==2.1.0
systemd-python==234
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
testresources==2.0.1
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.12.2
u-msgpack-python==2.3.0
ubuntu-pro-client==8001
ufbt==0.2.6
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.0.4
wrapt==1.16.0
wsaccel==0.6.3
zipp==1.0.0
zope.interface==5.4.0
